Implement lamp protocol
Lamp protocol is now implemented using basic functions and promises. The API is
very simple and straightforward.

  - protocol/ contains plug and discovery, probably the two most useful modules
  - protocol/data/ contains all the data parsing and serializing features (with
    tests)
  - utils/ contains error and lampId
    - error is an extensible error for custom exceptions
    - lampId creates a URL-friendly ID from the tuple (ip, port) and back

Tests are enabled using Istanbul for coverage reporting, Karma runner with
coverage module and a Babel plugin for instrumenting tests. The output is a
cobertura report.

Finally, the discovery is already implemented as a reducer (two actually) and it
fires a new dispatch every time a beacon is received

Config validation is currently not-implemented.

Change-Id: Icd20a509115cae08ed0d1de662505682961d3766