Bring support to the Android platform
An Android subproject has been initialized in order to port the app to
the platform. This required moving and reorganizing a bit the chrome
target in order to support side to side builds. Android supports the
wifi connection and discovery as in the desktop version. NFC will be
implemented in the future.

Important changes:
 - package.json: few new imports and new shortcuts for build steps
 - android/: the native code and build scripts
 - *.webpack.js: changes to allow side-to-side builds and correctly
   build a bundle for Android
 - src/protocol: native bridge protocol implementation
 - src/app.js: protocol initialization becomes platform specific

Change-Id: I3c8f1eb3e59cb34d471990e04209abd5142be31e