Change desktop target to Electron
Chrome apps will be discontinued soon [1]. With this commit, the desktop
platform becomes based on Electron [2], both for development and as the
final target.

Electron is a good choice for development because it supports sockets
out of the box, using Node's capabilities. Therefore the socket mocking
classes have been removed. Electron is also capable of loading DevTools
extensions, such as React DevTools, so there isn't any loss of
functionality in respect of Chrome.

README.md contains the instructions to develop both on Desktop and
Android platforms.

[1] http://blog.chromium.org/2016/08/from-chrome-apps-to-web.html
[2] http://electron.atom.io/

Change-Id: I7cd6e0c6c5d571cacdc04af78c4ba6abdd139362