Wait between calls to the Wi-Fi module
Sending two or more requests without pausing cause the module to close
the connection prematurely. This is a known behaviour: the WiFish module
acts as a bridge between a serial connection and a TCP connection.
Instead of refusing new connections that happen when a bridge between
the serial line and an existing TCP connection is in place, the module
accepts new connections and immediately close them without exchanging
any data.

Change-Id: If2bbd478514fe82fd0961f660a5cc00cb6a2747e