Create tests for emitter and collector
The tests are written in order to check that the math respects the
hand-calculated guidelines. They check whether the generated values
are in acceptable range and if they correspond to the math
specifications.

A line is changed in bugs.js in order to pass to the emitter
function an empty array of bugs.
This edit must be reviewed (ticket EPM-33)

Another line has been changed in README.md because the value
reported was not in the correct range (0-1 instead of 0-100)

Change-Id: Ia4894c1ffa6a0d9d28ee82abf37eef85771ab27d