Disable FlowType - ESLint integration (#12)
The integration has been disabled mainly because of Flow bugs. It's impossible to have a clean build without incurring some errors like "required module not found" or "missing property in object literal".

By merging this, the CI will no longer try to run Flow as a part of the build but it becomes a responsibility of the developer(s) to check for Flow errors (and all the usual things).

To make life easier, npm run flow is provided.