Edit Emitter to comply with robustness principle
Clamping operations have been removed, because the app
should always provide the function with acceptable data.
If inputs are not ok because they are out of range or because they
generate invalid results, an error is thrown.
Tests are modified with the same principle.
While editing for robustness tests have been modified for better
readability.

Change-Id: Ifeba81a1626352880a034dae35ba99fb39e44518